requirements.txt
python-kraken-sdk
